chore: 調整base entity id 型態

+++ src/main/java/com/cheng/linegroup/dao/MessageEventRecordRepository.java
@@ -3,5 +3,5 @@ package com.cheng.linegroup.dao;
 import com.cheng.linegroup.entity.MessageEventRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageEventRecordRepository extends JpaRepository<MessageEventRecord, Integer> {
+public interface MessageEventRecordRepository extends JpaRepository<MessageEventRecord, Long> {
 }
